refact(wecom): external contact tag group with agent id of pointer

+++ src/work/externalContact/tag/request/requestTagEditCorpTag.go
@@ -4,5 +4,5 @@ type RequestTagEditCorpTag struct {
 	ID      string `json:"id"`
 	Name    string `json:"name"`
 	Order   int    `json:"order"`
-	AgentID *int   `json:"agentid"`
+	AgentID *int64 `json:"agentid"`
 }
